requirements.txt
attrs==24.2.0
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.1
pillow==10.4.0
pydicom==2.4.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
zipp==3.20.2
